Move FileSettingsReaderTests
